requirements.txt
flet==0.26.dev3912
requests
python-dotenv